backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraudshield"
version = "0.5.0"
description = "Enterprise Fraud Risk Platform (Deterministic decisioning + ML scoring + agentic investigations)"
readme = "README.md"
requires-python = ">=3.11"

authors = [
  { name = "Ruslan Magana Vsevolodovna", email = "contact@ruslanmv.com" }
]

license = { text = "Proprietary" }

# ============================================================
# Core runtime dependencies (ALWAYS INSTALLED)
# ============================================================
dependencies = [
  # API / serving
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",

  # Core infra
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "sqlalchemy>=2.0",
  "pandas>=2.0",

  # ----------------------------------------------------------
  # Agentic runtime (core)
  # ----------------------------------------------------------
  "crewai>=0.30.0",

  # ----------------------------------------------------------
  # LLM providers (core – safe if unused)
  # ----------------------------------------------------------
  "openai>=1.30.0",
  "ibm-watsonx-ai>=1.0.0"
]

# ============================================================
# Optional extras
# ============================================================
[project.optional-dependencies]

# Ops / UI / orchestration helpers
ops = [
  "langchain>=0.2.0",
  "langchain_openai>=0.1.0",
  "streamlit>=1.31"
]

# ML lifecycle (training + registry demo)
ml = [
  "scikit-learn>=1.3",
  "joblib>=1.3"
]

# Cloud data platforms (connectors only)
cloud = [
  "google-cloud-bigquery>=3.0",
  "snowflake-connector-python>=3.0"
]

# Developer tooling
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6.0"
]

# ============================================================
# Packaging
# ============================================================
[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

# ============================================================
# Linting
# ============================================================
[tool.ruff]
line-length = 100
target-version = "py311"
